WORKING AFTER MESS

+++ F2/src/main/java/isel/sisinf/grp02/data_mappers/ICliente_InstitucionalMapper.java
@@ -2,5 +2,5 @@ package isel.sisinf.grp02.data_mappers;
 
 import isel.sisinf.grp02.orm.Cliente_Institucional;
 
-public interface ICliente_InstitucionalMapper extends IDataMapper<Cliente_Institucional, Long> {
+public interface ICliente_InstitucionalMapper extends IDataMapper<Cliente_Institucional, Integer> {
 }
